requirements.txt
fastapi
uvicorn[standard]
pydantic
flet
requests
mysql-connector-python